Refactor add_storage package into post process step
